Small fixes

+++ src/main/resources/static/main.js
@@ -22,25 +22,29 @@ function showTickets() {
         type: "GET",
         url: "/getTickets",
         success: function (data) {
-            // Formats the array into HTML
-            let output = "<h2>All tickets</h2><table class='table' id='tickets'>" +
-                "<tr><th>Movie</th><th>Number of tickets</th><th>First Name</th>" +
-                "<th>Last Name</th><th>Phone</th><th>E-mail</th><th></th><th></th></tr>"
-            for (let ticket of data) {
-                output += "<tr><td>"+ ticket.movie +"</td>" +
-                    "<td>"+ticket.amount+"</td>" +
-                    "<td>"+ticket.first_name+"</td>" +
-                    "<td>"+ticket.last_name+"</td>" +
-                    "<td>"+ticket.number+"</td>" +
-                    "<td>"+ticket.email+"</td>" +
+            let output = "";
+            // Checks if array is empty
+            if (data.length > 0) {
+                // Formats the array into HTML
+                output += "<h2>All tickets</h2><table class='table' id='tickets'>" +
+                    "<tr><th>Movie</th><th>Number of tickets</th><th>First Name</th>" +
+                    "<th>Last Name</th><th>Phone</th><th>E-mail</th><th></th><th></th></tr>"
+                for (let ticket of data) {
+                    output += "<tr><td>"+ ticket.movie +"</td>" +
+                        "<td>"+ticket.amount+"</td>" +
+                        "<td>"+ticket.first_name+"</td>" +
+                        "<td>"+ticket.last_name+"</td>" +
+                        "<td>"+ticket.number+"</td>" +
+                        "<td>"+ticket.email+"</td>" +
 
-                    // Adds buttons for editing and deleting tickets individually
-                    "<td><button class='btn btn-primary' onclick='editTicket("+ ticket.id +")'>Edit</button></td>" +
-                    "<td><button class='btn btn-danger' onclick='deleteTicket("+ ticket.id+")'>Delete</button></td>" +
-                    "</tr>"
+                        // Adds buttons for editing and deleting tickets individually
+                        "<td><button class='btn btn-primary' onclick='editTicket("+ ticket.id +")'>Edit</button></td>" +
+                        "<td><button class='btn btn-danger' onclick='deleteTicket("+ ticket.id+")'>Delete</button></td>" +
+                        "</tr>"
+                }
+                // Button for deleting all tickets
+                output += "</table><button class='btn btn-danger' onclick='deleteAllTickets()'>Delete all tickets</button>"
             }
-            // Button for deleting all tickets
-            output += "</table><button class='btn btn-danger' onclick='deleteAllTickets()'>Delete all tickets</button>"
             $("#tickets").html(output)
         }
     })
